requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.3
python-dotenv==1.0.1
qrcode==7.4.2
Pillow==10.2.0
SQLAlchemy==2.0.27
alembic==1.13.1
python-multipart==0.0.9
h11==0.14.0
pydantic-settings==2.1.0
aiosqlite==0.19.0
prometheus-client==0.20.0
